# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GBF ERP API
  version: "0.2.0"
  description: |
    通用贸易 MVP 后端 HTTP 接口（**精简说明**）。完整行为以 `src/routes` 与 Zod schema 为准。

    - 成功响应多为 `application/json`；错误体由全局错误处理统一包装。
    - 鉴权：除 `/health`、`/api/auth/login`、`/api/auth/refresh`（及开发态 `x-username`）外，一般需 `Authorization: Bearer <access_token>`。
servers:
  - url: http://localhost:3100
    description: 本地默认端口
...
paths:
...
  /api/sales-orders:
    get:
      tags: [orders]
      summary: 销售单列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    post:
      tags: [orders]
      summary: 创建销售单（草稿）
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SalesOrderCreate" }
      responses:
        "201":
          description: Created
  /api/sales-orders/{id}:
    get:
      tags: [orders]
      summary: 销售单详情（表头 + 明细行）
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: "{ order, items }"
  /api/sales-orders/{id}/action:
    post:
      tags: [orders]
      summary: 销售单动作
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderActionBody" }
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access token（`typ=access`）
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  schemas:
...
    OrderLine:
      type: object
      required: [productId, qty, price]
      properties:
        productId: { type: integer, minimum: 1 }
        qty: { type: number, minimum: 0 }
        price: { type: number, minimum: 0 }
...
    SalesOrderCreate:
      type: object
      required: [orderNo, customerId, items]
      properties:
        orderNo: { type: string, minLength: 3 }
        customerId: { type: integer, minimum: 1 }
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/OrderLine" }
    OrderActionBody:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [submit, approve, reject, void, reverse]
        comment: { type: string, maxLength: 200 }
...